Refactor Blackjack game logic and class names
The game logic for the Blackjack application is extensively revamped for better structure, clarity, and organization. Notably, several class names are altered for better context specificity such as "Table" to "BjTable" and "Game" to "Round". Additionally, a constant class has been created to make the game parameters more manageable and a dedicated simulation class is introduced to handle the execution and summary of game simulations.